Move all git operations to their own file
