+++ zet.go
@@ -402,100 +402,6 @@ func (z *Zet) ChangeDir(path string) error {
 	return nil
 }
 
-func (z *Zet) Pull() error {
-	// sanity check git remote
-	err := z.GitRemote()
-	if err != nil {
-		return err
-	}
-	err = z.ChangeDir(z.Path)
-	if err != nil {
-		return err
-	}
-	err = Z.Exec("git", "pull", "-q")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-func (z *Zet) Add() error {
-	err := z.ChangeDir(z.Path)
-	if err != nil {
-		return err
-	}
-	err = Z.Exec("git", "add", "-A", z.Path)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-func (z *Zet) Commit() error {
-	err := z.ChangeDir(z.Path)
-	if err != nil {
-		return err
-	}
-	err = Z.Exec("git", "commit", "-m", z.Title)
-	if err != nil {
-		return err
-	}
-	fmt.Printf("Committed %q\n", z.Title)
-	return nil
-}
-func (z *Zet) Push() error {
-	// sanity check git remote
-	err := z.GitRemote()
-	if err != nil {
-		return err
-	}
-	err = z.ChangeDir(z.Path)
-	if err != nil {
-		return err
-	}
-	err = Z.Exec("git", "push", "--quiet")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// PullAddCommitPush is a helper method which flows through a Git workflow
-// and is called often in Commands such as `create` and `edit`.
-func (z *Zet) PullAddCommitPush() error {
-	err := z.GetTitle()
-	if err != nil {
-		return err
-	}
-	err = z.Pull()
-	if err != nil {
-		return err
-	}
-	err = z.Add()
-	if err != nil {
-		return err
-	}
-	err = z.Commit()
-	if err != nil {
-		return err
-	}
-	err = z.Push()
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// GitRemote checks for the existence of a non-empty `git remote -v` response.
-func (z *Zet) GitRemote() error {
-	if os.Getenv("GIT_REMOTE") != "" {
-		return nil
-	}
-	s := Z.Out("git", "remote", "-v")
-	if s == "" {
-		return errors.New("no git remote found")
-	}
-	return nil
-}
-
 // Last inspects the Zet repo directories (Isosec folders) and returns the
 // most recent directory.
 func (z *Zet) Last() (string, error) {
